work on the running a downloaded c file
